fix vnni_dot saturating & fix sql position.
